// source/frontend/src/models/api/generated/ApiGen_Concepts_LeaseRenewal.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDateTime } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_Lease } from './ApiGen_Concepts_Lease';

// LINK: @backend/apimodels/Models/Concepts/Lease/LeaseRenewalModel.cs
export interface ApiGen_Concepts_LeaseRenewal extends ApiGen_Base_BaseAudit {
  id: number;
  leaseId: number;
  commencementDt: UtcIsoDateTime | null;
  expiryDt: UtcIsoDateTime | null;
  isExercised: boolean | null;
  renewalNote: string | null;
  lease: ApiGen_Concepts_Lease | null;
}